Removed Grid class due to not being a flexible enough approach for modelling the game.
Added Vector2D class to represent positions of Tetrominos, as well as to make it possible to easily compute rotations which was a shortcoming of the Grid class.

Added RotationMatrix2x2 class to compute the counterclockwise rotations of vectors represented by Vector2D.

Added Tetrominos enum to store the different types of tetrominos as well as some general functions.

Added tests and implementations for Vector2D class and RotationMatrix2x2 class.