BASIC GUI ADDED TO MAKE IT EASIER TO FIX BUGS.
NEED TO FIX ROTATION:
right now when the blocks are not at (0,0) any rotation applied will rotate it out of the screen. Need to rework tetromino class.
IDEAS:
- only track the location of the centre of the Tetro. Then can keep the pattern the same throughout movement and only use it when rotating, and when placing the block on the board.
i.e. when rotating only the pattern changes
i.e. when the block is placed on the board the locations of all the blocks can be calculated by adding the COM vector to each template vector.